Kobweb multimodule support
This change is a massive refactoring of the Kobweb Gradle
plugins, introducing a new library plugin (in addition to
the existing application plugin) as well as a core plugin
which both delegate to.

This commit also includes a sample project which showcases
the feature. See examples/multimodule for more details.

Fixes #85